refactor: Move auto search code into its own file
- Added an abort control and some "spam preventative" measures for not overloading the lyric providers.
